Separated tests for hrtimer
