+++ lib/hrtimer.js
@@ -1,12 +1,3 @@
-'use strict';
-
-/**
- * Measure latency for a load test.
- * (C) 2013 Alex Fernández.
- */
-
-
-const testing = require('testing');
 
 
 /**
@@ -14,7 +5,7 @@ const testing = require('testing');
  *	- delayMs: miliseconds to wait before calls. Can be fractional.
  *	- callback: function to call every time.
  */
-class HighResolutionTimer {
+export class HighResolutionTimer {
 
 	/**
 	 * Create a timer with the given delay.
@@ -70,44 +61,5 @@ class HighResolutionTimer {
 		}
 		return this
 	}
-}
-
-/**
- * Test a high resolution timer.
- */
-function testTimerStop(callback) {
-	const timer = new HighResolutionTimer(10, callback);
-	setImmediate(() => timer.stop())
-}
-
-function testTimerRun(callback) {
-	let run = 0
-	const timer = new HighResolutionTimer(100, () => run++)
-	setTimeout(() => {
-		testing.equals(run, 3, callback)
-		timer.stop()
-		testing.success(callback)
-	}, 250)
-}
-
-/**
- * Run package tests.
- */
-function test(callback) {
-	const tests = [
-		testTimerStop,
-		testTimerRun,
-	];
-	testing.run(tests, callback);
-}
-
-module.exports = {
-	HighResolutionTimer,
-	test,
-}
-
-// run tests if invoked directly
-if (__filename == process.argv[1]) {
-	test(testing.show);
 }
